% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{abbreviate_time}
\alias{abbreviate_time}
\title{Abbreviate time units}
\usage{
abbreviate_time(unit)
}
\arguments{
\item{unit}{Character. A character vector for a time units.}
}
\value{
Returns a character vector with one element.
}
\description{
Create abbreviations of time units for info blocks.
}
\examples{
\dontrun{
movedesign::abbreviate_time("years")
}

}
