% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{abbrv_unit}
\alias{abbrv_unit}
\title{Abbreviate units}
\usage{
abbrv_unit(unit)
}
\arguments{
\item{unit}{Character. A character vector for a unit.}
}
\value{
Returns a character vector with one element.
}
\description{
Create abbreviations of units.
}
\examples{
\dontrun{
movedesign::abbrv_unit("square kilometers")
}

}
