% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{fix_timeunits}
\alias{fix_timeunits}
\title{Fix values and units of time}
\usage{
fix_timeunits(value, unit)
}
\arguments{
\item{value}{numeric, integer. For example, 1.}

\item{unit}{character vector of time units. For example, "hours".}
}
\value{
A list with the corrected value and the corrected unit.
}
\description{
Correctly convert time values and units from ctmm outputs for blocks.
}
\examples{
\dontrun{
movedesign::fix_timeunits(1, "hours")
}

}
