% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fixrates}
\alias{fixrates}
\title{Fix rates of animal tracking devices.}
\format{
An object of class \code{"data.frame"},
with 40 rows and seven variables: \code{dti_notes}, \code{dti},
\code{highlight}, \code{frq}, and \code{frq_hrs}.

\describe{
  \item{dti_notes}{Sampling interval as text (e.g., "1 fix every month")}
  \item{dti}{Sampling interval in seconds}
  \item{frq}{Sampling frequency in seconds}
  \item{frq_hrs}{Sampling frequency in hours}
  \item{highlighted}{Highlight a group of commonly-used GPS fix rates}
  ...
}
}
\usage{
fixrates
}
\description{
Dataset with a list of fix rates typically available for animal tracking devices.
}
\keyword{datasets}
