% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gps_fixrate}
\alias{gps_fixrate}
\title{Fix rates of animal tracking devices.}
\format{
An object of class \code{"data.frame"},
with 40 rows and seven variables: \code{nu_notes}, \code{nu},
\code{highlight}, \code{choices}, \code{nu_hrs}, \code{freq}, and \code{freq_hrs}.

\describe{
  \item{nu_notes}{Sampling interval as text (e.g., "1 fix every month")}
  \item{nu}{Sampling interval in seconds}
  \item{nu_hrs}{Sampling interval in hours}
  \item{freq}{Sampling frequency in seconds}
  \item{freq_hrs}{Sampling frequency in hours}
  \item{highlight}{Highlight a group of commonly used fix rates}
  \item{choice}{Choose commonly used fix rates for GPS devices}
  ...
}
}
\usage{
gps_fixrate
}
\description{
Dataset with a list of fix rates typically available for animal tracking devices.
}
\keyword{datasets}
