% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gps_tradeoffs}
\alias{gps_tradeoffs}
\title{Dataset of GPS tradeoffs.}
\format{
Data frame
}
\usage{
gps_tradeoffs
}
\description{
Dataset showing how sample sizes changes with sampling duration and sampling interval.
}
\keyword{datasets}
