% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movedesign-package.R
\docType{package}
\name{movedesign-package}
\alias{movedesign}
\alias{movedesign-package}
\title{movedesign: Study design of movement ecology studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Shiny app to assist researchers in designing movement ecology studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecoisilva/movedesign}
  \item \url{https://ecoisilva.github.io/movedesign/}
  \item Report bugs at \url{https://github.com/ecoisilva/movedesign/issues}
}

}
\author{
\strong{Maintainer}: Inês i.simoes-silva@hzdr.de Silva

}
\keyword{internal}
