% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{movedesign_processed}
\alias{movedesign_processed}
\title{Create a \code{movedesign_processed} object for preprocessed results}
\usage{
movedesign_processed(input)
}
\arguments{
\item{input}{A list, containing preprocessed results and metadata.}
}
\value{
An object of class \code{movedesign_processed} and \code{movedesign}.
}
\description{
Creates an S3 object of class \code{movedesign_processed} to
store preprocessing steps or outputs during a \code{movedesign} workflow.
}
\seealso{
\code{\link[=md_run]{md_run()}}
}
\keyword{internal}
