% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movmods}
\alias{movmods}
\title{Table of movement processes.}
\format{
Data frame
}
\usage{
movmods
}
\description{
Table listing all current movement processes used
within the ctmm R package.
}
\keyword{datasets}
