% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{new_movedesign}
\alias{new_movedesign}
\title{Construct a \code{movedesign} S3 object}
\usage{
new_movedesign(x, subclass, ...)
}
\arguments{
\item{x}{The underlying data or object to be wrapped
(e.g., list, tibble, simulation object, output).}

\item{subclass}{Character string specifying the subclass
(e.g., \code{movedesign_input}, \code{movedesign_output}).}

\item{...}{Additional attributes to set on the object.}
}
\value{
An object of class \code{subclass} and \code{movedesign}.
}
\description{
A generic internal constructor for creating S3 objects
representing different stages of a "movedesign" workflow, such as
input, preprocessing, simulation, output, diagnostics, and plots.
}
\keyword{internal}
