% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_sims}
\alias{output_sims}
\title{Table of AKDE simulation outputs.}
\format{
Data frame
}
\usage{
output_sims
}
\description{
Table of outputs from AKDE simulations from Silva et al. (2021).
For further details, see \url{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.13786}
}
\keyword{datasets}
