% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{plot.movedesign_optimized}
\alias{plot.movedesign_optimized}
\title{Plot movedesign report outputs}
\usage{
\method{plot}{movedesign_optimized}(x, ...)
}
\arguments{
\item{x}{A \code{movedesign_optimized} object returned by
\code{md_optimized()} or similar.}

\item{...}{Unused}
}
\description{
S3 method for plotting a \code{movedesign_optimized} object.
Returns the precomputed ggplot stored in the object.
}
