% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{plot.movedesign_report}
\alias{plot.movedesign_report}
\title{Plot movedesign report outputs}
\usage{
\method{plot}{movedesign_report}(x, ...)
}
\arguments{
\item{x}{A \code{movedesign_report} object returned by
\code{md_optimize()} or similar.}

\item{...}{Unused}
}
\description{
S3 method for plotting a \code{movedesign_report} object.
Returns the precomputed ggplot stored in the object.
}
