% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plotting_hrsim}
\alias{plotting_hrsim}
\title{Plot home range with simulated data}
\usage{
plotting_hrsim(dat, datsim, ud, levels, show)
}
\description{
Plotting home range output from ctmm with simulation.
}
