% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{prepare_pars}
\alias{prepare_pars}
\title{Prepare movement model}
\usage{
prepare_pars(tau_p0, tau_p0_units, tau_v0, tau_v0_units, sigma0, sigma0_units)
}
\description{
Prepare parameters for movement data simulation
}
