% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{print.movedesign_check}
\alias{print.movedesign_check}
\title{Print method for \code{movedesign_check} objects}
\usage{
\method{print}{movedesign_check}(x, ...)
}
\arguments{
\item{x}{An object of class \code{movedesign_check}}

\item{...}{Unused}
}
\description{
Print method for \code{movedesign_check} objects
}
