% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{print.movedesign_input}
\alias{print.movedesign_input}
\title{Print method for \code{movedesign_input}}
\usage{
\method{print}{movedesign_input}(x, ...)
}
\arguments{
\item{x}{An object of class \code{movedesign_input}}

\item{...}{Additional arguments}
}
\description{
Print method for \code{movedesign_input}
}
