% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{print.movedesign_optimized}
\alias{print.movedesign_optimized}
\title{Print method for \code{movedesign_optimized} objects}
\usage{
\method{print}{movedesign_optimized}(
  x,
  verbose = TRUE,
  m = NULL,
  ci = 0.95,
  tol = 0.05,
  n_converge = 9,
  plot = TRUE,
  pal = c("#007d80", "#A12C3B"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{movedesign_optimized}}

\item{...}{Unused}
}
\description{
Print method for \code{movedesign_optimized} objects
}
