% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{print.movedesign_report}
\alias{print.movedesign_report}
\title{Print method for \code{movedesign_report} objects}
\usage{
\method{print}{movedesign_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{movedesign_report}}

\item{...}{Unused}
}
\description{
Print method for \code{movedesign_report} objects
}
