% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run movedesign R Shiny application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function called before the app runs. Only relevant for
programmatic usage.}

\item{options}{A named list passed to \code{shiny::runApp}.}

\item{enableBookmarking}{One of \code{url}, \code{server}, or \code{disable}. The
default, \code{NULL}, respects any previous call to \code{enableBookmarking}.}

\item{uiPattern}{A regular expression used to match request paths. The
request path must match the expression in full to be handled by the UI.}

\item{...}{arguments to pass to golem_opts.
See \code{?golem::get_golem_options} for more details.}
}
\value{
No return value. This function is called for its side effects.
}
\description{
Run movedesign R Shiny application
}
