% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sims_hrange}
\alias{sims_hrange}
\title{Home range error simulations.}
\format{
An object of class \code{"data.frame"}.
}
\usage{
sims_hrange
}
\description{
Expanded simulations from Silva et al. (2021).
}
\references{
Silva et al. (2021). Methods in Ecology and Evolution, 13(3), 534-544.
DOI: 10.1111/2041-210X.12559
}
\keyword{datasets}
