% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sims_speed}
\alias{sims_speed}
\title{Speed & distance error simulations.}
\format{
An object of class \code{"data.frame"}.
}
\usage{
sims_speed
}
\description{
Expanded simulations from Noonan et al. (2019) and Silva et al. (in prep).
}
\references{
Noonan et al. (2019).  Movement Ecology, 7(1), 1-15.
DOI: 10.1186/s40462-019-0177-1
}
\keyword{datasets}
