% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{simulate_gpsdecay}
\alias{simulate_gpsdecay}
\title{Simulate GPS battery life decay}
\usage{
simulate_gpsdecay(data, k0, yrange0, subset, minrate)
}
\arguments{
\item{data}{data.frame. A dataset with frequencies.}

\item{k0}{Numeric. The rate constant.}

\item{subset}{Character. Cut-off for frequencies}

\item{minrate}{Minimum frequency.}

\item{yrange}{Numeric. Value for the range of y, in y units.}
}
\description{
Simulate GPS battery life decay
}
