% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_input}
\alias{summary.movedesign_input}
\title{Summary method for \code{movedesign_input}}
\usage{
\method{summary}{movedesign_input}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_input}}

\item{...}{Additional arguments}
}
\description{
Summary method for \code{movedesign_input}
}
