% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_optimized}
\alias{summary.movedesign_optimized}
\title{Summary method for \code{movedesign_optimized} objects}
\usage{
\method{summary}{movedesign_optimized}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_optimized}}

\item{...}{Unused}
}
\description{
Summary method for \code{movedesign_optimized} objects
}
