% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_preprocess}
\alias{summary.movedesign_preprocess}
\title{Summary method for \code{movedesign_preprocess}}
\usage{
\method{summary}{movedesign_preprocess}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_preprocess}}

\item{...}{Additional arguments}
}
\description{
Summary method for \code{movedesign_preprocess}
}
