% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_processed}
\alias{summary.movedesign_processed}
\title{Summary method for \code{movedesign_processed}}
\usage{
\method{summary}{movedesign_processed}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_processed}}

\item{...}{Additional arguments}
}
\description{
Summary method for \code{movedesign_processed}
}
