% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_report}
\alias{summary.movedesign_report}
\title{Summary method for \code{movedesign_report} objects}
\usage{
\method{summary}{movedesign_report}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_report}}

\item{...}{Unused}
}
\description{
Summary method for \code{movedesign_report} objects
}
