% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{theme_movedesign}
\alias{theme_movedesign}
\title{movedesign ggplot2 custom theme}
\usage{
theme_movedesign(ft_size = 14)
}
\arguments{
\item{ft_size}{Base font size.}
}
\description{
Custom ggplot2 theme for movedesign plot outputs.
}
\author{
Ines Silva \email{i.simoes-silva@hzdr.de}
}
