% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gps_fixrate}
\alias{gps_fixrate}
\title{Dataset of GPS fix rates.}
\format{
A data frame with 40 observations and seven variables: \code{nu_notes}, \code{nu},
  \code{higlight}, \code{choices}, \code{nu_hrs}, \code{freq}, and \code{freq_hrs}.
}
\usage{
gps_fixrate
}
\description{
Dataset with a list of fix rates typically available for GPS tracking loggers.
}
\keyword{datasets}
