% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_main.R
\name{md_compare_preview}
\alias{md_compare_preview}
\title{Preview plot for movedesign workflow outputs (single replicate)}
\usage{
md_compare_preview(
  objs,
  n_resamples = NULL,
  error_threshold = 0.05,
  pal = c("#007d80", "#A12C3B"),
  ...
)
}
\arguments{
\item{objs}{A list of object of class \code{movedesign_preprocess}
(outputs of \code{\link[=md_run]{md_run()}}).}

\item{n_resamples}{Numeric. Must be a positive value. Defines how many
combinations are generated for each population sample size,
with each combination producing a new population-level estimate.}

\item{error_threshold}{Numeric. Error threshold (e.g. \code{0.05} for 5\%)
to display as a reference in the plot.}

\item{pal}{Character vector of two colors for within/outside threshold
(default: c("#007d80", "#A12C3B")).}

\item{...}{Additional arguments used internally.}
}
\value{
A ggplot object displaying relative error by population sample size,
with point estimate and confidence intervals for mean estimates,
and horizontal error threshold lines.
}
\description{
Generates a quick visualization of relative error for home range or
movement speed estimation from a single replicate of a movedesign
workflow.
The plot can display either the estimates from that replicate for a
random combination of individuals, or, when resampling is enabled,
summaries derived from repeated draws of individuals at each population
sample size (based on the specified number of resamples).

This functions shows preliminary outputs only based on the output of
\code{\link[=md_run]{md_run()}} (a \code{movedesign_preprocess} object) and should not be used to
evaluate study design by itself. Instead, users should run
\code{\link[=md_replicate]{md_replicate()}} and check for convergence with \code{\link[=md_check]{md_check()}}.
}
\details{
This plot summarizes a single replicate. Credible intervals and robust
study design conclusions generally require multiple replicates generated
with \code{\link[=md_replicate]{md_replicate()}}.
}
\examples{
if (interactive()) {
  inputA <- md_prepare(
    data = buffalo,
    models = models,
    species = "buffalo",
    n_individuals = 5,
    dur = list(value = 1, unit = "month"),
    dti = list(value = 1, unit = "day"),
    add_individual_variation = TRUE,
    grouped = TRUE,
    set_target = "hr",
    which_meta = "mean"
  )
  inputB <- md_prepare(
    data = buffalo,
    models = models,
    species = "buffalo",
    n_individuals = 5,
    dur = list(value = 10, unit = "days"),
    dti = list(value = 1, unit = "day"),
    add_individual_variation = TRUE,
    grouped = TRUE,
    set_target = "hr",
    which_meta = "mean"
  )

  outputA <- md_run(inputA)
  outputB <- md_run(inputB)
  md_compare_preview(list(outputA,
                          outputB), error_threshold = 0.05)
}

}
\seealso{
\code{\link[=md_run]{md_run()}}, \code{\link[=md_replicate]{md_replicate()}}
}
