% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_main.R
\name{md_plot_replicates}
\alias{md_plot_replicates}
\title{Plot replicate error estimates and confidence intervals}
\usage{
md_plot_replicates(obj, ci = 0.95)
}
\arguments{
\item{obj}{A movement design output object (see \code{\link[=md_replicate]{md_replicate()}}
or \code{\link[=md_stack]{md_stack()}}).}

\item{ci}{Numeric scalar between 0 and 1. The probability of the
credible interval (CI) to be estimated. Default to \code{0.95} (95\%).}
}
\value{
A list of class \code{movedesign_report} containing:
A list with two elements:
\itemize{
\item \code{p}: A \code{ggplot} object displaying the results from a single
replicate, showing individual error estimates and their confidence
intervals. It is equivalente to the output from \code{\link[=md_plot]{md_plot()}}.
\item \code{p.replicates}: A \code{ggplot} object summarizing mean errors
across all replicates, with aggregated estimates in the foreground
and individual replicates shown in lighter tones in the background.
}
}
\description{
This function generates two complementary visualizations of replicate
performance across different population sample sizes. It summarizes and
plots relative estimation errors for each type of analysis (e.g., home
range or speed estimation), distinguishing results that fall within or
outside a user-defined error threshold.
}
\examples{
if(interactive()) {
  obj <- md_replicate(...)
  
  plots <- md_plot_replicates(obj)
  # Display the plots:
  plots[[1]]
  plots[[2]]
}

}
\seealso{
\code{\link[=md_plot]{md_plot()}}
}
