% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movedesign-package.R
\docType{package}
\name{movedesign-package}
\alias{movedesign}
\alias{movedesign-package}
\title{movedesign: movedesign}
\description{
A shiny app to design movement ecology studies.
}
\author{
\strong{Maintainer}: Inês Silva \email{i.simoes-silva@hzdr.de}

}
\keyword{internal}
