% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movedesign-package.R
\docType{package}
\name{movedesign-package}
\alias{movedesign}
\alias{movedesign-package}
\title{movedesign: Study design of movement ecology studies}
\description{
The `movedesign` package contains an R shiny application that can be used
to assist researchers in designing movement ecology studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecoisilva/movedesign}
  \item \url{https://ecoisilva.github.io/movedesign/}
  \item Report bugs at \url{https://github.com/ecoisilva/movedesign/issues}
}

}
\author{
\strong{Maintainer}: Inês Silva \email{i.simoes-silva@hzdr.de} (\href{https://orcid.org/0000-0003-4850-6193}{ORCID})

}
\keyword{internal}
