% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_meta.R
\name{run_meta}
\alias{run_meta}
\title{Running \eqn{\chi^2}-IG hierarchical model meta-analyses}
\usage{
run_meta(
  rv,
  set_target = c("hr", "ctsd"),
  subpop = FALSE,
  trace = FALSE,
  iter_step = 2,
  .overwrite_seq = FALSE,
  .only_max_m = FALSE,
  .lists = NULL
)
}
\arguments{
\item{rv}{A list containing outputs, settings and data objects. Must not be NULL.}

\item{set_target}{Character. Research target: `"hr"` for home range or `"ctsd"` for speed & distance.}

\item{subpop}{Logical. If TRUE, will run meta-analyses with groups. Default is FALSE.}

\item{trace}{Logical. If TRUE, prints progress messages. Default is FALSE.}

\item{iter_step}{Numeric. The size of each iteration step. Default is 2.}

\item{.overwrite_seq}{Logical. If TRUE, overwrites sequence automatically to improve plot readability. Default is FALSE.}

\item{.only_max_m}{Logical. If TRUE, will only run the maximum number of individuals. Default is FALSE.}

\item{.lists}{A list containing already created meta inputs. Default is NULL.}
}
\value{
The outputs of the `run_meta_combinations` function for a single combination.
}
\description{
This function wraps around the `run_meta_combinations` function to run meta-analyses once (no combinations) for a quick evaluation.
}
