% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_check}
\alias{summary.movedesign_check}
\title{Summary method for \code{movedesign_check} objects}
\usage{
\method{summary}{movedesign_check}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_check}}

\item{...}{Unused}
}
\description{
Summary method for \code{movedesign_check} objects
}
