% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_main.R
\name{summary.movedesign_output}
\alias{summary.movedesign_output}
\title{Summary method for \code{movedesign_output}}
\usage{
\method{summary}{movedesign_output}(object, ...)
}
\arguments{
\item{object}{An object of class \code{movedesign_output}}

\item{...}{Additional arguments}
}
\description{
Summary method for \code{movedesign_output}
}
